test: rename folder name to exact same name as DB driver
